Lock SUSHI to 1.3.2
SUSHI 2.0.0 no longer supports the legacy project
structure used in this project. This commit adds a
fsh.ini file that tells the IG Publisher to used SUSHI
1.3.2 (no matter what the latest version of SUSHI
is).

This should be considered a temporary fix until the
project can be migrated to the supported project
structure.

See: https://fshschool.org/docs/sushi/migration/